src/requirements.txt
dateutils==0.6.12
mdutils==1.3.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
